Use compile/exec output rather than bool return
Otherwise the reducer could simply replace the "return x" with "return
true" and remove all the code.

This tool would be useful for compile errors and runtime panics. Both of
these can be determined by matching the output of the compilation and
execution of a program. Do that instead.

This shields us from falsely reducing code since removing the
interesting bits would either get rid of the string we want, or replace
it with another one.