Don't try the same source multiple times
For some reason, this was happening multiple times in the tests. Not in
the benchmark, though.

Likely some of the rules are slightly broken in the sense that they go
in circles. In any case, this is an easy change that will cover us from
those issues.

Test time drops from 3.5s to 2.5s.