Completely ignore go/types soft errors
When trying to reduce a program that triggers a cmd/compile internal
error, it's possible that it will reach the internal error before the
soft error (such as an unused variable).

After all, what we really need from go/types is the type information.
That can still be fetched even if there are soft errors. We were only
using them for debugging purposes and to avoid full round-trips to 'go
build'. But the latter is not a valid speedup if we want to be able to
work with internal compiler errors.

Fixes #13.