Make -run a shell cmd, remove main template magic
Templates weren't particularly useful. In the few cases where they are,
it's fairly simple to just add a main function with the code one needs.

Also drop the hard-coding of go build in the tool. It's unlikely that
we'll go lower-level than that, even with go tool compile. At least not
for the foreseeable future.

Fixes #18.